Modularize desktop main shell
Separate sidebar panel sizing, native window drag handling, and timeline chrome from the ClassicMainBody composition while preserving shell rendering and interactions.
