+++ apps/desktop/src/main/main-area-window-drag.ts
@@ -1,0 +1,180 @@
+import { isTauri } from "@tauri-apps/api/core";
+import { getCurrentWindow } from "@tauri-apps/api/window";
+import { type MouseEvent, type PointerEvent, useCallback, useRef } from "react";
+
+const MAIN_AREA_TOP_DRAG_HEIGHT_PX = 48;
+const MAIN_AREA_WINDOW_DRAG_THRESHOLD_PX = 5;
+
+type MainAreaWindowDragStart = {
+  pointerId: number;
+  clientX: number;
+  clientY: number;
+  dragging: boolean;
+};
+
+export function useMainAreaTopWindowDrag(enabled: boolean) {
+  const windowDragStartRef = useRef<MainAreaWindowDragStart | null>(null);
+  const suppressNextClickRef = useRef(false);
+
+  const handlePointerDown = useCallback(
+    (event: PointerEvent<HTMLDivElement>) => {
+      suppressNextClickRef.current = false;
+
+      if (
+        !enabled ||
+        event.button !== 0 ||
+        isInteractiveMainAreaDragTarget(event.target) ||
+        !isWithinMainAreaTopDragRegion(event)
+      ) {
+        windowDragStartRef.current = null;
+        return;
+      }
+
+      windowDragStartRef.current = {
+        pointerId: event.pointerId,
+        clientX: event.clientX,
+        clientY: event.clientY,
+        dragging: false,
+      };
+    },
+    [enabled],
+  );
+
+  const handlePointerMove = useCallback(
+    (event: PointerEvent<HTMLDivElement>) => {
+      const dragStart = windowDragStartRef.current;
+
+      if (
+        !dragStart ||
+        dragStart.dragging ||
+        dragStart.pointerId !== event.pointerId ||
+        !isMainAreaWindowDrag(dragStart, event)
+      ) {
+        return;
+      }
+
+      dragStart.dragging = true;
+      suppressNextClickRef.current = true;
+      event.preventDefault();
+
+      if (isTauri()) {
+        void getCurrentWindow()
+          .startDragging()
+          .catch(() => {});
+      }
+    },
+    [],
+  );
+
+  const handlePointerEnd = useCallback(
+    (event: PointerEvent<HTMLDivElement>) => {
+      const dragStart = windowDragStartRef.current;
+
+      if (!dragStart || dragStart.pointerId !== event.pointerId) {
+        return;
+      }
+
+      windowDragStartRef.current = null;
+    },
+    [],
+  );
+
+  const handleClickCapture = useCallback(
+    (event: MouseEvent<HTMLDivElement>) => {
+      if (!suppressNextClickRef.current) {
+        return;
+      }
+
+      suppressNextClickRef.current = false;
+      event.preventDefault();
+      event.stopPropagation();
+    },
+    [],
+  );
+
+  const handleDoubleClickCapture = useCallback(
+    (event: MouseEvent<HTMLDivElement>) => {
+      if (
+        !enabled ||
+        event.button !== 0 ||
+        isInteractiveMainAreaDragTarget(event.target) ||
+        isNativeWindowDragTarget(event.target) ||
+        !isWithinMainAreaTopDragRegion(event)
+      ) {
+        return;
+      }
+
+      event.preventDefault();
+      event.stopPropagation();
+
+      if (isTauri()) {
+        void getCurrentWindow()
+          .toggleMaximize()
+          .catch(() => {});
+      }
+    },
+    [enabled],
+  );
+
+  return {
+    onClickCapture: handleClickCapture,
+    onDoubleClickCapture: handleDoubleClickCapture,
+    onPointerDown: handlePointerDown,
+    onPointerEnd: handlePointerEnd,
+    onPointerMove: handlePointerMove,
+  };
+}
+
+function isWithinMainAreaTopDragRegion(
+  event: MouseEvent<HTMLDivElement> | PointerEvent<HTMLDivElement>,
+): boolean {
+  const rect = event.currentTarget.getBoundingClientRect();
+  const offsetY = event.clientY - rect.top;
+
+  return offsetY >= 0 && offsetY < MAIN_AREA_TOP_DRAG_HEIGHT_PX;
+}
+
+function isInteractiveMainAreaDragTarget(target: EventTarget | null): boolean {
+  if (!(target instanceof Element)) {
+    return false;
+  }
+
+  return Boolean(
+    target.closest(
+      [
+        "a",
+        "button",
+        "input",
+        "select",
+        "textarea",
+        "[contenteditable='true']",
+        "[role='button']",
+        "[role='textbox']",
+      ].join(","),
+    ),
+  );
+}
+
+function isNativeWindowDragTarget(target: EventTarget | null): boolean {
+  if (!(target instanceof Element)) {
+    return false;
+  }
+
+  return (
+    target.hasAttribute("data-tauri-drag-region") &&
+    target.getAttribute("data-tauri-drag-region") !== "false"
+  );
+}
+
+function isMainAreaWindowDrag(
+  start: { clientX: number; clientY: number },
+  current: { clientX: number; clientY: number },
+): boolean {
+  const deltaX = current.clientX - start.clientX;
+  const deltaY = current.clientY - start.clientY;
+
+  return (
+    deltaX * deltaX + deltaY * deltaY >=
+    MAIN_AREA_WINDOW_DRAG_THRESHOLD_PX * MAIN_AREA_WINDOW_DRAG_THRESHOLD_PX
+  );
+}
